EmailSender integration tests
