+++ src/integrationTest/resources/application.properties
@@ -12,8 +12,12 @@ clients.error-notifications.username=username
 
 flyway.skip-migrations=false
 
-spring.datasource.driver-class-name=org.testcontainers.jdbc.ContainerDatabaseDriver
-spring.datasource.url=jdbc:tc:postgresql:9.6://localhost/blob_router
+#spring.datasource.driver-class-name=org.testcontainers.jdbc.ContainerDatabaseDriver
+#spring.datasource.url=jdbc:tc:postgresql:9.6://localhost/blob_router
+spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.url=jdbc:h2:mem:testdb
+spring.datasource.username=sa
+spring.datasource.password=
 
 storage-blob-processing-delay-in-minutes=0
 public-key-der-file=signing/test_public_key.der
@@ -31,7 +35,9 @@ scheduling.task.handle-rejected-files.cron=0/10 * * * * *
 
 scheduling.task.delete-rejected-files.enabled=false
 
-spring.mail.host=smpth.localhost
-spring.mail.username=username
-spring.mail.password=password
+spring.mail.host=localhost
+# GreenMail port
+spring.mail.port=3025
+spring.mail.username=test_username
+spring.mail.password=test_password
 spring.mail.test-connection=false
